refactor (update package) : MenuGroup 도메인 패키지 내부 구조 변경
package
- bo -> application
- controller -> api
- model -> domain

class name
- MenuGroupBo -> MenuGroupService
