+++ src/main/java/camp/nextstep/edu/kitchenpos/menugroup/dao/MenuGroupDao.java
@@ -1,6 +1,6 @@
 package camp.nextstep.edu.kitchenpos.menugroup.dao;
 
-import camp.nextstep.edu.kitchenpos.menugroup.model.MenuGroup;
+import camp.nextstep.edu.kitchenpos.menugroup.domain.MenuGroup;
 
 import java.util.List;
 import java.util.Optional;
